gigachat_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
cachetools